+++ CHANGELOG.md
@@ -10,7 +10,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Support for mysql via sqlalchemy
-- SQL server interface
+- SQL server parameters added to config, env, and Dockerfiles
+
+### Changed
+
+- Database schema has been expanded to support more metadata fields
+- The `DELETE` endpoint's response has been modified since a batch deletion is used
+- Workflow database has been renamed from `results` to `workflow`
+- Default database filename is changed from `results.db` to `dispatch.db`
 
 ## [0.98.0] - 2022-05-15